requirements.txt
# Core Flask application
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
python-socketio==5.8.0

# Azure Cognitive Services
azure-cognitiveservices-speech==1.34.0
azure-identity==1.15.0

# OpenAI and LLM
openai==1.3.0
langchain==0.0.350
langchain-openai==0.0.2

# Vector Database and Embeddings
pinecone-client==2.2.4
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# Speech Processing
SpeechRecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.11

# Data Processing
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0

# Database
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
mysqlclient==2.1.1

# Security and Compliance
cryptography==41.0.4
python-dotenv==1.0.0

# Monitoring and Logging
loguru==0.7.2
prometheus-client==0.17.1

# Testing
pytest==7.4.2
pytest-cov==4.1.0

# Development
black==23.9.1
flake8==6.1.0
mypy==1.6.1

# Utilities
requests==2.31.0
python-dateutil==2.8.2
Pillow==10.0.1 